+++ src/context/SearchContext.tsx
@@ -1,22 +1,7 @@
-import { createContext, Key, ReactNode, useReducer } from 'react';
-import { DateRange } from '@lib/classes';
-import { SearchFilterSet, SearchResultCandidate } from '@lib/types';
+import { createContext, ReactNode, useReducer } from 'react';
+import { DateRange, SearchFilterSet } from '@lib/classes';
+import { SearchResultCandidate } from '@lib/types';
 import { additionalFilterKeys } from '@lib/utils';
-
-const emptyFilterSet: SearchFilterSet = {
-	positions: {
-		departments: [],
-		jobs: [],
-	},
-	skills: [],
-	jobDates: new DateRange(),
-	unions: [],
-	locations: [],
-	experienceLevels: [],
-	genderIdentities: [],
-	racialIdentities: [],
-	personalIdentities: [],
-};
 
 interface SearchState {
 	filters: {
@@ -41,8 +26,9 @@ interface SearchAction {
 		skills?: string[];
 		locations?: string[];
 		filter?: {
-			name: string;
-			value: string | string[] | Key[];
+			key: string;
+			value: any;
+			position?: 'departments' | 'jobs';
 		};
 		jobDates?: DateRange;
 		filterSet?: SearchFilterSet;
@@ -52,22 +38,27 @@ interface SearchAction {
 	};
 }
 
-const initialSearchState: SearchState = {
+/**
+ * Returns the initial state for the search context.
+ *
+ * @return {SearchState} The initial search state with default values for filters, search activity, active filters, saved search, and results.
+ */
+const getInitialSearchState = (): SearchState => ({
 	filters: {
 		name: '',
-		filterSet: emptyFilterSet,
+		filterSet: new SearchFilterSet(),
 	},
 	searchActive: false,
 	additionalFiltersActive: [],
 	savedSearch: {
 		id: 0,
-		filterSet: emptyFilterSet,
+		filterSet: new SearchFilterSet(),
 	},
 	results: [],
-};
+});
 
 export const SearchContext = createContext({
-	search: initialSearchState,
+	search: getInitialSearchState(),
 	searchDispatch: ({}: SearchAction) => {},
 });
 
@@ -77,104 +68,58 @@ function searchContextReducer(state: SearchState, action: SearchAction): SearchS
 			return {
 				...state,
 				filters: {
-					...initialSearchState.filters,
+					...getInitialSearchState().filters,
 					name: action.payload.name ? action.payload.name : '',
 				},
 				// Clear all other filters and set the main search controls to inactive
 				searchActive: false,
 			};
 
-		case 'SET_DEPARTMENT':
-			if (!action.payload?.departments) return state;
+		case 'SET_POSITIONS': {
+			if (
+				!action.payload?.filter ||
+				(action.payload.filter.key !== 'departments' && action.payload.filter.key !== 'jobs')
+			)
+				return state;
+
+			const filterSet = state.filters.filterSet;
+			filterSet.setPositions(action.payload.filter.key, action.payload.filter.value);
 
 			return {
 				...state,
 				filters: {
 					...state.filters,
-					filterSet: {
-						...state.filters.filterSet,
-						positions: {
-							departments: action.payload.departments,
-							// Clear jobs
-							jobs: [],
-						},
-						// Clear skills
-						skills: [],
-					},
+					filterSet,
 				},
 				searchActive: true,
 			};
+		}
 
-		case 'SET_JOBS':
-			if (!action.payload?.jobs) return state;
-
-			return {
-				...state,
-				filters: {
-					...state.filters,
-					filterSet: {
-						...state.filters.filterSet,
-						positions: {
-							...state.filters.filterSet.positions,
-							jobs: action.payload.jobs,
-						},
-					},
-				},
-				searchActive: true,
-			};
-
-		case 'SET_SKILLS':
-			if (!action.payload?.skills) return state;
-
-			return {
-				...state,
-				filters: {
-					...state.filters,
-					filterSet: {
-						...state.filters.filterSet,
-						skills: action.payload.skills,
-					},
-				},
-				searchActive: true,
-			};
-
-		case 'SET_FILTER':
+		case 'SET_FILTER': {
 			if (!action.payload?.filter) return state;
+			const { key, value } = action.payload.filter;
+
+			const filterSet = state.filters.filterSet;
+			filterSet.set(key, value);
 
 			return {
 				...state,
 				filters: {
 					...state.filters,
-					filterSet: {
-						...state.filters.filterSet,
-						[action.payload.filter.name]: action.payload.filter.value,
-					},
+					filterSet,
 				},
 				searchActive: true,
 			};
+		}
 
-		case 'SET_JOB_DATES':
-			if (!action.payload?.jobDates) return state;
-
-			return {
-				...state,
-				filters: {
-					...state.filters,
-					filterSet: {
-						...state.filters.filterSet,
-						jobDates: action.payload.jobDates,
-					},
-				},
-				searchActive: true,
-			};
-
-		case 'SET_ADDITIONAL_FILTERS_ACTIVE':
+		case 'SET_ADDITIONAL_FILTERS_ACTIVE': {
 			if (!action.payload?.additionalFiltersActive) return state;
 
 			return {
 				...state,
 				additionalFiltersActive: action.payload.additionalFiltersActive,
 			};
+		}
 
 		case 'SET_SAVED_SEARCH_FILTERS': {
 			const {
@@ -232,7 +177,7 @@ function searchContextReducer(state: SearchState, action: SearchAction): SearchS
 			};
 
 		case 'RESET_SEARCH_FILTERS':
-			return initialSearchState;
+			return getInitialSearchState();
 
 		default:
 			return state;
@@ -240,7 +185,7 @@ function searchContextReducer(state: SearchState, action: SearchAction): SearchS
 }
 
 export const SearchContextProvider = ({ children }: { children: ReactNode }) => {
-	const [search, searchDispatch] = useReducer(searchContextReducer, initialSearchState);
+	const [search, searchDispatch] = useReducer(searchContextReducer, getInitialSearchState());
 
 	return (
 		<SearchContext.Provider value={{ search, searchDispatch }}>{children}</SearchContext.Provider>